renamed to BestBookDetailsParser based on it's working and locale fixed
